Allow dirs to exist and proceed with workflow
This change will allow the workflow to run several times on the
same persistent cluster. The disadvantage is that currently
some of the logs of the Parsl apps will simply be appended
to as more and more workflows run -> the result is that you
can have logs from multiple workflow runs appending into the
same Parsl app stdout and stderr. For most folks, this will
only be a minor nuance when it comes to debugging.